#261 [mod] : 이미지 멀티파트 화 확장 함수 변경
